backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]>=0.23.0
motor==3.3.2
pymongo==4.6.0
python-dotenv==1.0.0
pydantic==2.5.0
bcrypt==4.0.1
passlib==1.7.4
httpx>=0.25.0,<0.28.0
